Collapse select into one field and add formik
