--- ui/components/atoms/A015-Select/stories.js
@@ -1,77 +1,0 @@
-import React from 'react'
-import { storiesOf } from '@storybook/react'
-import { withNotes } from '@storybook/addon-notes'
-import { StateDecorator, Store } from '@sambego/storybook-state'
-
-// Import MenuItem for example children
-import { MenuItem } from '@material-ui/core'
-
-import Select from '.'
-import notes from './readme.md'
-
-const Countries = [
-  {
-    name: 'New Zealand'
-  },
-  {
-    name: 'Fiji'
-  },
-  {
-    name: 'Samoa'
-  },
-  {
-    name: 'Tonga'
-  }
-]
-
-// TODO DL Currently this is not working :(
-const store = new Store({
-  country: ''
-})
-
-storiesOf('Atoms|A015 - Select', module)
-  .addDecorator(StateDecorator(store))
-  .addDecorator(withNotes)
-  .add(
-    'Default',
-    () => (
-      <Select
-        id='1'
-        type='text'
-        name='Select Country'
-        value={store.get('country')}
-        onChange={event => store.set({ country: event.target.value })}
-      >
-        {Countries.map(option => (
-          <MenuItem key={option.name} value={option.name}>
-            {option.name}
-          </MenuItem>
-        ))}
-      </Select>
-    ),
-    {
-      notes: { markdown: notes }
-    }
-  )
-  .add(
-    'HelpText & Error',
-    () => (
-      <Select
-        id='1'
-        type='text'
-        name='Select Country'
-        value={store.get('country')}
-        onChange={event => store.set({ country: event.target.value })}
-        error
-      >
-        {Countries.map(option => (
-          <MenuItem key={option.name} value={option.name}>
-            {option.name}
-          </MenuItem>
-        ))}
-      </Select>
-    ),
-    {
-      notes: { markdown: notes }
-    }
-  )
