--- ui/components/atoms/A016-SelectLabel/index.js
@@ -1,59 +1,0 @@
-import React from 'react'
-import createPropTypes from 'json-schema-prop-types'
-import { withStyles } from '@material-ui/core/styles'
-
-import { TextField } from '@material-ui/core'
-
-import schema from './schema'
-import styles from './styles'
-
-const A016 = props => {
-  const {
-    id,
-    value,
-    required,
-    disabled,
-    name,
-    onChange,
-    readOnly,
-    error,
-    classes,
-    label,
-    children
-  } = props
-
-  const shownHelperText = error ? 'Select an option' : null
-  return (
-    <TextField
-      id={id}
-      value={value}
-      required={required}
-      disabled={disabled}
-      name={name}
-      onChange={onChange}
-      InputProps={{
-        readOnly: readOnly
-      }}
-      error={error}
-      FormHelperTextProps={{
-        error: error
-      }}
-      SelectProps={{
-        MenuProps: {
-          className: classes.menu
-        }
-      }}
-      helperText={shownHelperText}
-      className={styles.formField}
-      label={label}
-      margin='normal'
-      select
-    >
-      {children}
-    </TextField>
-  )
-}
-
-A016.propTypes = createPropTypes(schema)
-
-export default withStyles(styles)(A016)
