--- ui/components/atoms/A016-SelectLabel/readme.md
@@ -1,6 +1,0 @@
-# TextArea
-
-## API
-
-**name**: The name attribute is used to reference elements in a JavaScript, or to reference form data after a form is submitted.
-
