--- ui/components/atoms/A016-SelectLabel/styles.js
@@ -1,10 +1,0 @@
-export default theme => ({
-  formField: {
-    marginLeft: theme.spacing.unit,
-    marginRight: theme.spacing.unit,
-    width: 200
-  },
-  menu: {
-    width: 200
-  }
-})
